Change USER and install sudo
Looking through the original Dockerfile used to create snipe-it,
a "docker" user has been created, and various "chown's" have been
performed for it, so it seems sensible to use "docker" and not
"snipe".

First error to look at is:

    + '[' -f /var/lib/snipeit/ssl/snipeit-ssl.crt -a -f /var/lib/snipeit/ssl/snipeit-ssl.key ']'
    + a2dismod ssl
    Could not remove /etc/apache2/mods-enabled/ssl.load: Permission denied

Digging around the working container, everything under /etc/apache2 is
owned by "root".  Easiest way to let "a2dismod" is to install sudo,
and configure the docker for it.