Get supervisord to actually start apache
  Action '-DFOREGROUND' failed.
  The Apache error log may have more information.
  2020-08-12 23:28:30,265 INFO exited: apache (exit status 1; not expected)

It appears that apache needs to be started as root, in order for
'-DFOREGROUND' to work.  As 'supervisord' is running as 'docker',
it can't start apache correctly.

After much playing round with the supervisord "user=" option, and
trying to find out more about how 'FOREGROUND' actually works (it's
amazingly undocumented), I resorted to using 'sudo' again