Fix apache DocumentRoot
  D'oh. If you're seeing this, your DocumentRoot is set incorrectly. It should be set to the public directory.

apache is starting, and serving pages.  Just not the right ones.

The message is coming from /var/www/html/index.html, but what
should being served is /var/www/html/public/index.php

Tracked it down to a "VirtualHost" setting that is using
port 80, and not 8080 as we've told apache to listen on